refactor: move subcollection operation to rhcConnection DAO
It also renames it so that the name is consistent with the other
subcollection functions of other daos.
