+++ dao/interfaces.go
@@ -19,8 +19,6 @@ type SourceDao interface {
 	Tenant() *int64
 	NameExistsInCurrentTenant(name string) bool
 	GetByIdWithPreload(id *int64, preloads ...string) (*m.Source, error)
-	// GetRelatedRhcConnectionsToId gets all the related connections to the given source id.
-	GetRelatedRhcConnectionsToId(sourceId *int64, limit, offset int, filters []util.Filter) ([]m.RhcConnection, int64, error)
 }
 
 type ApplicationDao interface {
@@ -111,4 +109,6 @@ type RhcConnectionDao interface {
 	Delete(id *int64) error
 	// GetRelatedSourcesToId gets all the sources that are related to a given rhcConnection id.
 	GetRelatedSourcesToId(rhcConnectionId *int64, limit, offset int, filters []util.Filter) ([]m.Source, int64, error)
+	// ListForSource gets all the related connections to the given source id.
+	ListForSource(sourceId *int64, limit, offset int, filters []util.Filter) ([]m.RhcConnection, int64, error)
 }
