+++ dao/mock_daos.go
@@ -9,8 +9,7 @@ import (
 )
 
 type MockSourceDao struct {
-	Sources               []m.Source
-	RelatedRhcConnections []m.RhcConnection
+	Sources []m.Source
 }
 
 type MockApplicationDao struct {
@@ -34,8 +33,9 @@ type MockMetaDataDao struct {
 }
 
 type MockRhcConnectionDao struct {
-	RhcConnections []m.RhcConnection
-	RelatedSources []m.Source
+	RhcConnections        []m.RhcConnection
+	RelatedRhcConnections []m.RhcConnection
+	RelatedSources        []m.Source
 }
 
 func (src *MockSourceDao) SubCollectionList(primaryCollection interface{}, limit, offset int, filters []util.Filter) ([]m.Source, int64, error) {
@@ -120,12 +120,6 @@ func (src *MockSourceDao) GetByIdWithPreload(id *int64, preloads ...string) (*m.
 	}
 
 	return nil, fmt.Errorf("source not found")
-}
-
-func (m *MockSourceDao) GetRelatedRhcConnectionsToId(sourceId *int64, limit, offset int, filters []util.Filter) ([]m.RhcConnection, int64, error) {
-	count := int64(len(m.RelatedRhcConnections))
-
-	return m.RelatedRhcConnections, count, nil
 }
 
 func (a *MockApplicationTypeDao) List(limit int, offset int, filters []util.Filter) ([]m.ApplicationType, int64, error) {
@@ -441,4 +435,10 @@ func (m *MockRhcConnectionDao) GetRelatedSourcesToId(id *int64, limit, offset in
 	count := int64(len(m.RelatedSources))
 
 	return m.RelatedSources, count, nil
+}
+
+func (m *MockRhcConnectionDao) ListForSource(sourceId *int64, limit, offset int, filters []util.Filter) ([]m.RhcConnection, int64, error) {
+	count := int64(len(m.RelatedRhcConnections))
+
+	return m.RelatedRhcConnections, count, nil
 }
